requirements.txt
numpy
pandas
pyarrow
Pillow
tqdm
rootutils
projectaria-tools[all]
opencv-python
open3d